pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["petriknovis <dpetrik@novis.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<=3.10"
streamlit = "^1.32.2"
geopandas = "^0.14.3"
pandas = "^1.5.2"
jupyter = "^1.0.0"
pyarrow = "^10.0.1"
tqdm = "^4.64.1"
hsfs = "3.4.8"
plotly = "^5.11.0"
requests = "^2.28.1"
scikit-learn = "^1.3.1"
xgboost = "^2.0.0"
optuna = "^3.4.0"
hopsworks = "^3.4.2"
python-dotenv = "^1.0.0"
pydeck = "^0.8.0"
discordwebhook = "^1.0.3"
fire = "^0.5.0"
lightgbm = "^4.1.0"
comet-ml = "^3.35.3"
retry = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
